+++ CHANGES.rst
@@ -4,7 +4,7 @@
 * First major steps on armechelle
 * Coadding; including docs; and tests
 * Now requiring astropy >= v1.3
-
+* raw_input handling for Python 3
 
 0.7 (2017-02-07)
 ----------------